Rename jobs package to job, move GA statistics to metrics, minor edits
