+++ main.go
@@ -37,7 +37,7 @@ import (
 
 	"go.uber.org/zap"
 
-	"github.com/Arriven/db1000n/src/jobs"
+	"github.com/Arriven/db1000n/src/job"
 	"github.com/Arriven/db1000n/src/runner"
 	"github.com/Arriven/db1000n/src/runner/config"
 	"github.com/Arriven/db1000n/src/utils"
@@ -51,7 +51,7 @@ func main() {
 	log.Printf("DB1000n [Version: %s][PID=%d]\n", ota.Version, os.Getpid())
 
 	runnerConfigOptions := runner.NewConfigOptionsWithFlags()
-	jobsGlobalConfig := jobs.NewGlobalConfigWithFlags()
+	jobsGlobalConfig := job.NewGlobalConfigWithFlags()
 	otaConfig := ota.NewConfigWithFlags()
 	countryCheckerConfig := utils.NewCountryCheckerConfigWithFlags()
 	updaterMode, destinationPath := config.NewUpdaterOptionsWithFlags()
@@ -107,11 +107,10 @@ func newZapLogger(debug bool) (*zap.Logger, error) {
 }
 
 func setUpPprof(pprof string, debug bool) {
-	if debug && pprof == "" {
+	switch {
+	case debug && pprof == "":
 		pprof = ":8080"
-	}
-
-	if pprof == "" {
+	case pprof == "":
 		return
 	}
 
@@ -122,9 +121,7 @@ func setUpPprof(pprof string, debug bool) {
 	mux.Handle("/debug/pprof/symbol", http.HandlerFunc(pprofhttp.Symbol))
 	mux.Handle("/debug/pprof/trace", http.HandlerFunc(pprofhttp.Trace))
 
-	go func() {
-		log.Println(http.ListenAndServe(pprof, mux))
-	}()
+	go log.Println(http.ListenAndServe(pprof, mux))
 }
 
 func cancelOnSignal(cancel context.CancelFunc) {
